+++ changelog.md
@@ -1,3 +1,6 @@
+- add a Renderers module, now have a class method to pull memoized instances of renderers
+- reorganize a bit, moved "modules" and renamed them to "formatters"
+
 # 0.3.1
 - fix the header_with_id
 - add an Extensions module for rendering modules that need extensions